[Refactor] #643 - NotificationBell.vue를 common 디렉토리로 이동
- NavBar.vue import 경로 수정
