+++ frontend/src/components/common/NotificationBell.vue
